# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchEventOutcome": {
        "properties": {
          "client_event_id": {
            "title": "Client Event Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "client_event_id",
          "success"
        ],
        "title": "BatchEventOutcome",
        "type": "object"
      },
      "BlockedOn": {
        "description": "Which check a BLOCKED or NEEDS_OVERRIDE state is blocked on.",
        "enum": [
          "LOCATION",
          "SKU",
          "QTY"
        ],
        "title": "BlockedOn",
        "type": "string"
      },
...
      "EventResult": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "blocked_on": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockedOn"
              },
              {
                "type": "null"
              }
            ]
          },
          "effects": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Effects",
            "type": "array"
          },
          "last_qty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Qty"
          },
          "line_id": {
            "format": "uuid",
            "title": "Line Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PickState"
          }
        },
        "required": [
          "line_id",
          "state",
          "attempts",
          "blocked_on",
          "last_qty",
          "effects"
        ],
        "title": "EventResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentReason": {
        "description": "Reasons an operator can raise an exception on a pick line.",
        "enum": [
          "short",
          "empty_location",
          "damaged"
        ],
        "title": "IncidentReason",
        "type": "string"
      },
      "IncidentResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pick_line_id": {
            "format": "uuid",
            "title": "Pick Line Id",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/IncidentReason"
          },
          "reported_qty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Qty"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pick_line_id",
          "session_id",
          "reason",
          "reported_qty",
          "resolved",
          "created_at"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "badge_code": {
            "title": "Badge Code",
            "type": "string"
          },
          "pin": {
            "title": "Pin",
            "type": "string"
          }
        },
        "required": [
          "badge_code",
          "pin"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "operator_id": {
            "format": "uuid",
            "title": "Operator Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/OperatorRole"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "operator_id",
          "full_name",
          "role"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "OperatorProfile": {
        "properties": {
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "operator_id": {
            "format": "uuid",
            "title": "Operator Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/OperatorRole"
          }
        },
        "required": [
          "operator_id",
          "full_name",
          "role"
        ],
        "title": "OperatorProfile",
        "type": "object"
      },
      "OperatorRole": {
        "description": "Who an operator is for access-control purposes (Phase 4).",
        "enum": [
          "operator",
          "supervisor"
        ],
        "title": "OperatorRole",
        "type": "string"
      },
...
      "PickState": {
        "description": "States a single pick line can be in.",
        "enum": [
          "PENDING",
          "AWAITING_LOCATION",
          "AWAITING_SCAN",
          "BLOCKED",
          "NEEDS_OVERRIDE",
          "AWAITING_QTY",
          "SHORT_PENDING",
          "DONE",
          "EXCEPTED"
        ],
        "title": "PickState",
        "type": "string"
      },
...
      "SubmitEventBatchRequest": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/SubmitEventRequest"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "SubmitEventBatchRequest",
        "type": "object"
      },
      "SubmitEventBatchResponse": {
        "properties": {
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/BatchEventOutcome"
            },
            "title": "Outcomes",
            "type": "array"
          }
        },
        "required": [
          "outcomes"
        ],
        "title": "SubmitEventBatchResponse",
        "type": "object"
      },
      "SubmitEventRequest": {
        "properties": {
          "client_event_id": {
            "title": "Client Event Id",
            "type": "string"
          },
          "event": {
            "additionalProperties": true,
            "title": "Event",
            "type": "object"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "pick_line_id": {
            "format": "uuid",
            "title": "Pick Line Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "client_event_id",
          "pick_line_id",
          "session_id",
          "event"
        ],
        "title": "SubmitEventRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "macenplast-voice-picking",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Lets a client confirm a stored token is still valid.",
        "operationId": "get_me_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorProfile"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/events": {
      "post": {
        "operationId": "submit_event_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Submit Event",
        "tags": [
          "events"
        ]
      }
    },
    "/events/batch": {
      "post": {
        "operationId": "submit_event_batch_events_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitEventBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitEventBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Submit Event Batch",
        "tags": [
          "events"
        ]
      }
    },
...
    "/incidents": {
      "get": {
        "operationId": "list_incidents_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "resolved",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resolved"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response List Incidents Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Incidents",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/resolve": {
      "post": {
        "operationId": "resolve_incident_incidents__incident_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Resolve Incident",
        "tags": [
          "incidents"
        ]
      }
    },
...
  }
}